Preliminary Cleanup Commit
Still need to organize my CSS, but most of the issues have been addressed. There are a couple finicky ones I have to come back to.  With the size of this project, it's become more apparent how organizing my CSS as I go will be helpful. This was pretty arduous when I came back to clean everything up.